requirements.txt
torch~=2.5.1
transformers~=4.48.0
Pillow~=11.1.0
opencv-python~=4.10.0.84
tesseract
flask~=3.1.0
fastapi~=0.115.6
uvicorn
pytesseract~=0.3.13
PaddleOCR
camelot-py
pdf2image~=1.17.0
numpy~=1.26.4
werkzeug~=3.1.3
itsdangerous~=2.2.0
pydantic~=2.10.4
requests~=2.32.3
filelock~=3.16.1
click~=8.1.8
openpyxl~=3.1.5
pandas~=2.2.3
python-dateutil~=2.9.0.post0
chardet~=5.2.0
scikit-learn~=1.6.0
tqdm~=4.64.0
pypdf~=3.17.4
six~=1.17.0
pyclipper~=1.3.0.post6
shapely~=2.0.6
networkx~=3.4.2
pdf2image~=1.17.0
openai~=1.59.3
pdfplumber~=0.11.5
beartype~=0.19.0
httpx~=0.28.1
colorama~=0.4.6
lxml~=5.3.0
beautifulsoup4~=4.12.3
soupsieve~=2.6
h11~=0.14.0
cryptography~=44.0.0
cffi~=1.17.1
pip~=22.3.1
attrs~=24.3.0
distro~=1.9.0
wheel~=0.38.4
setuptools~=65.5.1
pytz~=2024.2
sniffio~=1.3.1
exceptiongroup~=1.2.2
idna~=3.10
PyYAML~=6.0.2
zipp~=3.21.0
urllib3~=2.3.0
httpcore~=1.0.7
certifi~=2024.12.14
jiter~=0.8.2
scipy~=1.14.1
jinja2~=3.1.5
Cython~=3.0.11
regex~=2024.11.6
threadpoolctl~=3.5.0
packaging~=24.2
matplotlib~=3.10.0
mpmath~=1.3.0
sympy~=1.13.1
fsspec~=2024.12.0
MarkupSafe~=3.0.2
joblib~=1.4.2
anyio~=4.7.0
tomli~=2.2.1
blinker~=1.9.0
pyparsing~=3.2.1
contourpy~=1.3.1
fonttools~=4.55.3
pycparser~=2.22
cycler~=0.12.1
kiwisolver~=1.4.8
tokenizers~=0.21.0
mock~=4.0.3
safetensors~=0.5.0
xgboost~=2.1.3
hanziconv~=0.3.2
nltk~=3.9.1
onnxruntime~=1.20.1